+++ lib/boring_backup/stores/store.rb
@@ -1,5 +1,13 @@
 module BoringBackup::Stores
   class Store
+    def name
+      self.class.name.split("::").last.downcase
+    end
+
+    def description
+      name
+    end
+
     private
 
     def config